Simplified closure creation MakeCls.
In the original compiler there was a case in Closure
named MakeCls ((x, t), clos, e) that essentially bound x (with
function type t) to the closure clos and then evaluated e in this
environment. I simplified this to MakeClos (clos), so that the
'old' MakeCls ((x, t), clos, e) becomes Let ((x, t), MakeCls (clos), e).
Thus we can reuse some of the code we already use for the usual Let.