Simplified interm. repr. of primitive operations.
Inspired by the sources of Caml Light (which can
be found at <http://caml.inria.fr/caml-light/>),
I introduced a Primitive case in the intermediate
languages KNormal, Closure, and GC. The possible
primitives are listed in prim.ml. Since all these
primitives are treated almost exactly the same in
every one of the intermediate phases, this saves
quite a lot of typing.