Reimplemented Less Equal (<) for Float values.
This can not really be considered a satisfactory
solution. Since LE (and also, Eq) is polymorphic, I add
a type (initialiser to Type.gentyp ()) to the LE node
and unify it with the type of the operands. Then in KNormal
I use this type to desambiguate the operator according the the
types of its operands.